refactor: 쿠키 유틸리티를 서버/브라우저 환경별로 분리
- cookie.ts를 cookie.server.ts와 cookie.browser.ts로 분리
- 서버 사이드 API 라우트에서 cookie.server 임포트
- 클라이언트 컴포넌트에서 cookie.browser 임포트
- 'server-only' 지시문을 통해 서버 전용 유틸리티 보장
