+++ apps/web/src/app/api/comments/route.ts
@@ -1,6 +1,6 @@
 import CommentService from '@/services/comment.service';
 import { NextRequest, NextResponse } from 'next/server';
-import { verifyRequestCookies } from '@/utils/cookie';
+import { verifyRequestCookies } from '@/utils/cookie.server';
 import { ResponseError } from '@moneed/utils';
 import { ERROR_MSG, SUCCESS_MSG } from '@/constants/message';
