+++ apps/web/src/app/api/posts/[postId]/route.ts
@@ -1,5 +1,5 @@
 import PostService from '@/services/post.service';
-import { verifyRequestCookies } from '@/utils/cookie';
+import { verifyRequestCookies } from '@/utils/cookie.server';
 import { NextRequest, NextResponse } from 'next/server';
 import { UpdatePostRequest } from '@/types/post';
 import { ResponseError } from '@moneed/utils';
