+++ apps/web/src/app/api/users/me/comments/route.ts
@@ -1,4 +1,4 @@
-import { verifyRequestCookies } from '@/utils/cookie';
+import { verifyRequestCookies } from '@/utils/cookie.server';
 import CommentService from '@/services/comment.service';
 import { NextResponse } from 'next/server';
 import { ResponseError } from '@moneed/utils';
