+++ apps/web/src/components/Comment/Comment.tsx
@@ -5,7 +5,7 @@ import { PrimaryDropdown, PrimaryDropdownProps } from '@/components/Dropdown';
 import Image from 'next/image';
 import { Comment as TComment } from '@/types/post';
 import DateFormatter from '../Dateformatter';
-import { getCookie } from '@/utils/cookie';
+import { getCookie } from '@/utils/cookie.browser';
 import { TOKEN_KEY } from '@moneed/auth';
 import { decodeJwt } from 'jose';
 import type { TokenPayload } from '@moneed/auth';
